# src/docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  dtos.DeviceTemplateBootStageDto:
    properties:
      action:
        description: "Action for this boot stage.\nCan be: File, CheckPowerSwitch,
          EmergencyPowerOff,\nPowerOff, EmergencyPowerOn, PowerOn,\nCheckManagement\n\nFor
          File action:\n\tA stage can only be marked complete if all files have\n\tbeen
          downloaded by the device via TFTP or DHCP,\n\tafter which the next step
          can be loaded."
        type: string
      description:
        description: Description of boot stage template
        type: string
      files:
        description: Files slice of files for boot stage
        items:
          $ref: '#/definitions/dtos.DeviceTemplateBootStageFileDto'
        type: array
      name:
        description: Name of boot stage template
        type: string
    type: object
  dtos.DeviceTemplateBootStageFileDto:
    properties:
      existingFileName:
        description: |-
          ExistingFileName file name is a real full file path with name on the disk.
          This path is relative from app directory
        type: string
      virtualFileName:
        description: VirtualFileName virtual file name is relative from /<mac-address>/
        type: string
    type: object
  dtos.DeviceTemplateControlDto:
    properties:
      emergency:
        description: 'Emergency how to control device power in case of emergency.
          As example: POE(For Rpi4), IPMI, ILO or PowerSwitch'
        type: string
      nextBoot:
        description: |-
          NextBoot how to change next boot device. As example: IPMI, ILO or NONE.
          For example, NONE is used for Rpi4, we control next boot by u-boot files in boot stages.
        type: string
      power:
        description: 'Power how to control device power. As example: POE(For Rpi4),
          IPMI, ILO or PowerSwitch'
        type: string
    type: object
  dtos.DeviceTemplateDto:
    properties:
      control:
        $ref: '#/definitions/dtos.DeviceTemplateControlDto'
        description: Control describes how we control the device
      cpucount:
        description: CPUCount count of cpus
        type: integer
      cpumodel:
        description: CPUModel model of cpu
        type: string
      description:
        description: Description template description
        type: string
      discBootStages:
        description: DiscBootStages slice of boot stage templates for disk boot
        items:
          $ref: '#/definitions/dtos.DeviceTemplateBootStageDto'
        type: array
      manufacturer:
        description: Manufacturer device manufacturer
        type: string
      model:
        description: Model device model
        type: string
      name:
        description: Name template name
        type: string
      netBootStages:
        description: NetBootStages slice of boot stage templates for net boot
        items:
          $ref: '#/definitions/dtos.DeviceTemplateBootStageDto'
        type: array
      networkInterfaces:
        description: NetworkInterfaces slice of device network interfaces
        items:
          $ref: '#/definitions/dtos.DeviceTemplateNetworkDto'
        type: array
      ram:
        description: RAM the amount of RAM in GB
        type: integer
      usbbootStages:
        description: USBBootStages slice of boot stage templates for usb boot
        items:
          $ref: '#/definitions/dtos.DeviceTemplateBootStageDto'
        type: array
    type: object
  dtos.DeviceTemplateNetworkDto:
    properties:
      management:
        description: Management only one network interface can be mark as management
        type: boolean
      name:
        description: Name of network interface. This field is unique within device
          template network interfaces
        type: string
      netBoot:
        description: NetBoot flags whether the interface can be loaded over the network
        type: boolean
      poein:
        description: POEIn only one network interface can be mark as POEIn
        type: boolean
    type: object
  dtos.DtoType:
    type: object
...
  dtos.PaginatedListDto:
    properties:
      items:
        description: Items - array of DTO items
        items:
          $ref: '#/definitions/dtos.DtoType'
        type: array
      page:
        description: Page - page number
        type: integer
      size:
        description: Size - page size
        type: integer
      total:
        description: Total - total number of items
        type: integer
    type: object
  dtos.ResponseDataDto:
    properties:
      data:
        description: Data - any data to send
      status:
        $ref: '#/definitions/dtos.ResponseStatusDto'
        description: Status - response status structure
    type: object
...
  dtos.ResponseStatusDto:
    properties:
      code:
        description: Code - 0 if OK, otherwise 1
        type: integer
      message:
        description: Message - response message
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API supporter
    url: http://www.swagger.io/support
  description: Description of specifications
  license:
    name: license(Mandatory)
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Rack of labs API
  version: version(1.0)
paths:
...
  /template/device/:
    get:
      consumes:
      - application/json
      parameters:
      - description: Order by field
        in: path
        name: orderBy
        type: string
      - description: '''asc'' or ''desc'' for ascending or descending order'
        in: path
        name: orderDirection
        type: string
      - description: searchable value in entity
        in: path
        name: search
        type: string
      - description: page number
        in: path
        name: page
        type: integer
      - description: number of entities per page
        in: path
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dtos.ResponseDataDto'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/dtos.PaginatedListDto'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/dtos.DeviceTemplateDto'
                        type: array
                    type: object
              type: object
      summary: Gets paginated list of device templates
      tags:
      - device template
  /template/device/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: device template name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dtos.ResponseDataDto'
            - properties:
                data:
                  $ref: '#/definitions/dtos.DeviceTemplateDto'
              type: object
      summary: Gets device template by its name
      tags:
      - device template
swagger: "2.0"